Modified friends api, added custom permissions
User has to be in gallery/image owner's friends to interact with
the corresponding gallery or image.
Friend requests will not be used, but instead we are going
to have a simple follow system.
A user cannot follow oneself, serializer field was filtered
to exclude request.user from available friends to follow
Comment model added.