Major changes, read on
Added an nginx server to act as a reverse proxy.
It checks the uri in question and forwards the request to the django
container if it's not a request for static or media files.
If it's a media/static request it forwards it to one of the nginx
containers using basic round robin load balancing.
There is room for a lot of optimization regarding the loads balancing
part, but it will be implemented once we have integrated zookeeper to
our project and completed the frontend.
Finally, u can access the webserver at :8080, which is the only exposed
port to the host.